refactor: code quality improvements and bug fixes
Bug Fixes:
- Fix duplicate 'Schema X not found' error message in handlers.rs
- Fix sql_only returning 200 on errors (now returns 400)
- Fix shortestPath path variable 'p' not resolved in SELECT
- Fix with_cross_table CTE reference bug

Dead Code Removal (~400+ lines):
- Remove unused VarSource::Parameter variant
- Simplify SchemaEntity to unit variant (fields unused)
- Remove dead current_cte_name field from ScopeContext
- Remove dead extract_final_filters method and trait
- Remove dead try_build_join_based_plan method
- Remove dead path_variable field and methods from multi_type_vlp_joins

Test Stability:
- Fix flaky tests caused by GLOBAL_SCHEMAS race condition
- Change try_write() to blocking loop for test isolation
- Add #[serial(global_schema)] for cross-file test serialization

Code Quality:
- Add Default impl for GraphJoinInference
- Move test-only imports into #[cfg(test)] modules
- Remove unused imports in query_planner

All 832 tests passing, 0 clippy errors
