+++ src/query_planner/analyzer/mod.rs
@@ -90,7 +90,8 @@ mod duplicate_scans_removing;
 pub mod errors;
 mod filter_tagging;
 mod graph_context;
-mod graph_join_inference;
+pub mod graph_join;
+pub use graph_join as graph_join_inference;
 mod graph_traversal_planning;
 mod group_by_building;
 mod plan_sanitization;
