+++ src/query_planner/analyzer/type_inference.rs
@@ -82,21 +82,16 @@ use crate::{
             analyzer_pass::{AnalyzerPass, AnalyzerResult},
             errors::AnalyzerError,
         },
-        logical_expr::{Direction, LogicalExpr},
+        logical_expr::LogicalExpr,
         logical_plan::{GraphNode, GraphRel, LogicalPlan, ViewScan},
         plan_ctx::PlanCtx,
         transformed::Transformed,
     },
 };
 
-/// Maximum number of relationship types that can be inferred when unspecified.
-/// Prevents excessive UNION ALL expansion that would generate extremely large queries.
-/// If inference would result in more types, user must specify explicit type(s).
-///
-/// Limit of 5: Each inferred type creates a separate CTE branch in the query.
-/// For complex patterns with multiple inferences, this can combinatorially explode.
-/// Example: 5 edge types × 3 node labels = 15 CTE branches = very large SQL.
-const MAX_INFERRED_TYPES: usize = 5;
+// Note: MAX_INFERRED_TYPES constant was removed as dead code.
+// The default value (5) is now set directly in plan_builder.rs
+// via `max_inferred_types.unwrap_or(5)` and documented in plan_ctx.
 
 pub struct TypeInference;
 
@@ -897,301 +892,6 @@ impl TypeInference {
             );
         }
     }
-
-    fn infer_edge_types(
-        &self,
-        current_types: &Option<Vec<String>>,
-        left_connection: &str,
-        right_connection: &str,
-        _direction: &Direction,
-        plan_ctx: &mut PlanCtx,
-        graph_schema: &GraphSchema,
-    ) -> AnalyzerResult<Option<Vec<String>>> {
-        // Strategy 1: If types already specified, use them
-        if current_types.is_some() {
-            log::debug!(
-                "🔗 TypeInference: Edge types already specified: {:?}",
-                current_types
-            );
-            return Ok(current_types.clone());
-        }
-
-        let rel_schemas = graph_schema.get_relationships_schemas();
-
-        // Strategy 2: If schema has only one relationship, use it
-        // Extract unique type names (excluding composite keys)
-        let all_rel_types: Vec<String> = rel_schemas
-            .keys()
-            .filter_map(|k| {
-                // Extract type name from composite key "TYPE::FROM::TO" or simple "TYPE"
-                k.split("::").next().map(|s| s.to_string())
-            })
-            .collect::<std::collections::HashSet<_>>() // Deduplicate
-            .into_iter()
-            .collect();
-
-        if all_rel_types.len() == 1 {
-            log::info!(
-                "🔗 TypeInference: Only one edge type in schema, using: {}",
-                all_rel_types[0]
-            );
-            return Ok(Some(vec![all_rel_types[0].clone()]));
-        }
-
-        // Strategy 3: Try to get node labels from plan_ctx
-        let left_label = plan_ctx
-            .get_table_ctx(left_connection)
-            .ok()
-            .and_then(|ctx| ctx.get_label_opt().map(|s| s.to_string()));
-
-        let right_label = plan_ctx
-            .get_table_ctx(right_connection)
-            .ok()
-            .and_then(|ctx| ctx.get_label_opt().map(|s| s.to_string()));
-
-        log::debug!(
-            "🔗 TypeInference: Inferring edge type for '{}' ({:?}) -> '{}' ({:?})",
-            left_connection,
-            left_label,
-            right_connection,
-            right_label
-        );
-
-        // If both labels unknown and multiple relationships exist, can't infer
-        if left_label.is_none() && right_label.is_none() {
-            log::debug!(
-                "🔗 TypeInference: Both nodes untyped and schema has {} relationships, cannot infer",
-                rel_schemas.len()
-            );
-            return Ok(None);
-        }
-
-        // Strategy 3: Filter relationships by node type compatibility
-        // Supports polymorphic relationships with $any wildcards and label_values
-        let matching_types: Vec<String> = rel_schemas
-            .iter()
-            .filter(|(_, rel_schema)| {
-                // Helper: Check if query label matches schema (supports $any and label_values)
-                let matches_from = |query_label: &Option<String>| -> bool {
-                    query_label
-                        .as_ref()
-                        .map(|l| {
-                            // Direct match
-                            if l == &rel_schema.from_node || &rel_schema.from_node == "$any" {
-                                return true;
-                            }
-                            // Check polymorphic from_label_values
-                            if let Some(values) = &rel_schema.from_label_values {
-                                return values.contains(l);
-                            }
-                            false
-                        })
-                        .unwrap_or(true) // None means any label OK
-                };
-
-                let matches_to = |query_label: &Option<String>| -> bool {
-                    query_label
-                        .as_ref()
-                        .map(|l| {
-                            // Direct match
-                            if l == &rel_schema.to_node || &rel_schema.to_node == "$any" {
-                                return true;
-                            }
-                            // Check polymorphic to_label_values
-                            if let Some(values) = &rel_schema.to_label_values {
-                                return values.contains(l);
-                            }
-                            false
-                        })
-                        .unwrap_or(true) // None means any label OK
-                };
-
-                // CRITICAL: Parser already normalized left=from, right=to
-                // Direction field is only for display, not for logical structure!
-                // Always check: left_label with from_node, right_label with to_node
-                matches_from(&left_label) && matches_to(&right_label)
-            })
-            .map(|(type_name, _)| type_name.clone())
-            .collect();
-
-        if matching_types.is_empty() {
-            log::debug!(
-                "🔗 TypeInference: No relationships found connecting {:?} -> {:?}",
-                left_label,
-                right_label
-            );
-            return Ok(None);
-        }
-
-        // Check if too many types would result in excessive UNION branches
-        if matching_types.len() > plan_ctx.max_inferred_types {
-            let types_preview: Vec<_> = matching_types.iter().take(5).cloned().collect();
-            let types_str = if matching_types.len() > 5 {
-                format!("{}, ...", types_preview.join(", "))
-            } else {
-                matching_types.join(", ")
-            };
-
-            return Err(AnalyzerError::InvalidPlan(format!(
-                "Too many matching relationship types ({}) for {:?}->{:?}: [{}]. Max allowed is {}. Please specify explicit relationship type(s).",
-                matching_types.len(), left_label, right_label, types_str, plan_ctx.max_inferred_types
-            )));
-        }
-
-        if matching_types.len() == 1 {
-            log::info!(
-                "🔗 TypeInference: Inferred edge type '{}' from node labels {:?} -> {:?}",
-                matching_types[0],
-                left_label,
-                right_label
-            );
-        } else {
-            log::info!(
-                "🔗 TypeInference: Multiple matching types {:?} for {:?} -> {:?}, will expand to UNION",
-                matching_types, left_label, right_label
-            );
-        }
-
-        Ok(Some(matching_types))
-    }
-
-    /// Get existing node label or infer from edge type(s).
-    ///
-    /// Returns: Some(label) if found/inferred, None if couldn't infer.
-    /// Side effect: Updates plan_ctx if label was inferred.
-    fn get_or_infer_node_label(
-        &self,
-        node_alias: &str,
-        edge_types: &Option<Vec<String>>,
-        is_from_side: bool,
-        plan_ctx: &mut PlanCtx,
-        graph_schema: &GraphSchema,
-    ) -> AnalyzerResult<Option<String>> {
-        // Try to get existing label from plan_ctx
-        if let Ok(table_ctx) = plan_ctx.get_table_ctx(node_alias) {
-            if let Some(label) = table_ctx.get_label_opt() {
-                log::debug!(
-                    "🏷️ TypeInference: '{}' already has label: {}",
-                    node_alias,
-                    label
-                );
-                return Ok(Some(label));
-            }
-        }
-
-        // No existing label - try to infer from edge type(s)
-        let edge_types = match edge_types {
-            Some(types) => types,
-            None => {
-                log::debug!(
-                    "🏷️ TypeInference: '{}' has no label and relationship has no types",
-                    node_alias
-                );
-                return Ok(None);
-            }
-        };
-
-        if edge_types.is_empty() {
-            log::debug!(
-                "🏷️ TypeInference: '{}' has no label and relationship has empty types",
-                node_alias
-            );
-            return Ok(None);
-        }
-
-        // For multiple relationship types, we can't infer reliably
-        // (different types might have different node labels)
-        if edge_types.len() > 1 {
-            log::debug!(
-                "🏷️ TypeInference: '{}' has no label and relationship has multiple types {:?} - can't infer",
-                node_alias,
-                edge_types
-            );
-            return Ok(None);
-        }
-
-        let rel_type = &edge_types[0];
-
-        log::debug!(
-            "🏷️ TypeInference: Trying to look up relationship schema for type '{}'",
-            rel_type
-        );
-
-        // Look up relationship schema
-        let rel_schema = graph_schema
-            .get_relationships_schema_opt(rel_type)
-            .ok_or_else(|| {
-                log::error!(
-                    "🏷️ TypeInference: Relationship type '{}' NOT FOUND in schema!",
-                    rel_type
-                );
-                AnalyzerError::InvalidPlan(format!(
-                    "Relationship type '{}' not found in schema",
-                    rel_type
-                ))
-            })?;
-
-        log::debug!(
-            "🏷️ TypeInference: Found relationship schema for '{}': from_node={}, to_node={}",
-            rel_type,
-            rel_schema.from_node,
-            rel_schema.to_node
-        );
-
-        // Infer label from schema based on whether node is on from or to side
-        let inferred_label = if is_from_side {
-            &rel_schema.from_node
-        } else {
-            &rel_schema.to_node
-        };
-
-        log::info!(
-            "🏷️ TypeInference: Inferred '{}' label as '{}' from {} side of '{}' relationship",
-            node_alias,
-            inferred_label,
-            if is_from_side { "from" } else { "to" },
-            rel_type
-        );
-
-        // Update or CREATE TableCtx with inferred label
-        // This is the elegant solution: carry known constraints forward
-        // If the node was unlabeled during parsing, we now know its label from the relationship schema
-        if let Some(table_ctx) = plan_ctx.get_mut_table_ctx_opt(node_alias) {
-            // TableCtx exists - update it with inferred label
-            table_ctx.set_labels(Some(vec![inferred_label.clone()]));
-            log::info!(
-                "🏷️ TypeInference: UPDATED plan_ctx['{}'] = '{}' (from {} side of {})",
-                node_alias,
-                inferred_label,
-                if is_from_side { "from" } else { "to" },
-                rel_type
-            );
-        } else {
-            // TableCtx doesn't exist yet - CREATE it with inferred label
-            // This happens for unlabeled intermediate nodes like (b) in (a)-[r1]->(b)-[r2]->(c)
-            use crate::query_planner::plan_ctx::TableCtx;
-            plan_ctx.insert_table_ctx(
-                node_alias.to_string(),
-                TableCtx::build(
-                    node_alias.to_string(),
-                    Some(vec![inferred_label.clone()]),
-                    vec![], // No properties yet (will be added by property resolver)
-                    false,  // Not a relationship
-                    false,  // Not explicitly named by user (inferred)
-                ),
-            );
-            log::info!(
-                "🏷️ TypeInference: CREATED plan_ctx['{}'] = '{}' (from {} side of {})",
-                node_alias,
-                inferred_label,
-                if is_from_side { "from" } else { "to" },
-                rel_type
-            );
-        }
-
-        Ok(Some(inferred_label.clone()))
-    }
-
     /// Infer the label/type of elements being unwound from an UNWIND expression.
     ///
     /// Strategy:
